Swift 6 update - small changes

+++ ToyFlightSimulator Shared/Graphics/Shaders/TFSCommon.h
@@ -140,8 +140,8 @@ typedef struct
     matrix_float4x4 projection_matrix;
     matrix_float4x4 projection_matrix_inverse;
     matrix_float4x4 view_matrix;
-    uint framebuffer_width;
-    uint framebuffer_height;
+    uint64_t framebuffer_width;
+    uint64_t framebuffer_height;
 
     // Per Mesh frameData
     matrix_float4x4 temple_modelview_matrix;
